Make SynthVoice.cpp JUCE independent + Organize files
